+++ src/lib/services/wasm/txs/index.ts
@@ -1,4 +1,4 @@
 export * from "./api";
 export * from "./gql";
-// export * from "./lcd";
+export * from "./lcd";
 export * from "./types";