+++ src/lib/types/move/abi.ts
@@ -1,57 +1,15 @@
+import type { HexAddr } from "../addrs";
 import type { Nullable } from "../common";
-import type { SnakeToCamelCaseNested } from "../converter";
-import type { Pagination } from "../rest";
-
-export enum UpgradePolicy {
-  ARBITRARY = "ARBITRARY",
-  COMPATIBLE = "COMPATIBLE",
-  IMMUTABLE = "IMMUTABLE",
-}
 
 export type Visibility = "public" | "friend" | "private" | "script";
 
-// TODO: revisit address type later
-export interface ABIModule {
-  address: string;
-  name: string;
-  functions: ABIFunction[];
-}
-
-// TODO: revisit moduleAddress type later
-interface ABIFunction {
-  method: "query" | "tx";
-  moduleAddress: string;
-  moduleName: string;
-  functionName: string;
-  typeArgsLength: number;
-  argsTypes: string[];
-}
-
-/* response */
-// TODO: change address type to HexAddr after figuring out how to correctly infer NominalType intersection
-export interface ResponseModule {
-  address: string;
-  module_name: string;
-  abi: string;
-  raw_bytes: string;
-  upgrade_policy: UpgradePolicy;
-}
-
-export interface ResponseModules {
-  modules: ResponseModule[];
-  pagination: Pagination;
-}
-
-// TODO: change address type to HexAddr after figuring out how to correctly infer NominalType intersection
-export interface ResponseABI {
-  address: string;
+export interface ModuleAbi {
+  address: HexAddr;
   name: string;
   friends: string[];
   exposed_functions: ExposedFunction[];
   structs: Struct[];
 }
-
-export type InternalModule = SnakeToCamelCaseNested<ResponseModule>;
 
 export interface ExposedFunction {
   name: string;